Code cleanup, bug fixes, and comments.
Release mode perf statistics: 1.75 million games simulated per second, on a single thread, i7 Intel processor.

+++ Fast-Poker-Sim/src/HoldEm.cpp
@@ -63,7 +63,7 @@ void HoldEm::Stop()
 	this->session.complete = true;
 }
 
-bool HoldEm::Update(int& gamesPlayed)
+bool HoldEm::Update(uint64_t& gamesPlayed)
 {
 	assert(gamesPlayed >= 0);
 	assert(this->isPlaying == true);
